prep for point free refactor
