+++ average.js
@@ -1,0 +1,11 @@
+/**
+ * Average two numbers
+ *
+ * @param {Number} a First operand
+ * @param {Number} b Second operand
+ *
+ * @returns {Number} Average of a and b
+ */
+const average = (a, b) => (a + b) / 2;
+
+module.exports = { average };
