+++ index.test.js
@@ -1,39 +1,12 @@
 const expect = require('chai').expect;
 
 const {
-  bucket,
   median,
   octile,
   quantile,
   quartile,
   tercile,
 } = require('./');
-
-describe('bucket', () => {
-  describe('evenly', () => {
-    it('', () => {
-      expect(bucket(2, [5, 6, 7, 8, 1, 2, 3, 4]))
-      .to.eql([[1, 2, 3, 4], [5, 6, 7, 8]]);
-    });
-
-    it('sorts numerically, not alphabetically', () => {
-      expect(bucket(2, [2, 10])) // numeric should return [[2], [10]], alpha [[10], [2]]
-      .to.eql([[2], [10]]);
-    });
-  });
-
-  describe('unevenly', () => {
-    it('', () => {
-      expect(bucket(3, [5, 6, 7, 8, 1, 2, 3, 4]))
-      .to.eql([[1, 2, 3], [4, 5, 6], [7, 8]]);
-    });
-
-    it('with duplicates on the split boundary', () => {
-      expect(bucket(3, [5, 6, 5, 3, 1, 2, 3, 4]))
-      .to.eql([[1, 2, 3], [3, 4, 5], [5, 6]]);
-    });
-  });
-});
 
 describe('quantile', () => {
   it('evenly', () => {
