+++ package.json
@@ -21,7 +21,8 @@
     "mocha": "^6.1.4"
   },
   "dependencies": {
-    "lodash.curry": "^4.1.1"
+    "lodash.curry": "^4.1.1",
+    "lodash.flow": "^3.5.0"
   },
   "scripts": {
     "test": "mocha \"*.test.js\""
