+++ sort.js
@@ -1,0 +1,10 @@
+/**
+ * Default numerical sort
+ *
+ * @param {Array} arr The array to sort
+ *
+ * @returns {Array} The sorted array
+ */
+const sort = arr => arr.sort((a, b) => a - b);
+
+module.exports = { sort };
